pyproject.toml
[project]
name = "data-tagger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "json-repair>=0.47.6",
    "lingua-language-detector>=2.1.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "vllm>=0.9.1",
]
